Moving templates to dir
