+++ module/data/actor/templates/character.mjs
@@ -1,5 +1,5 @@
-import { E20 } from "../../helpers/config.mjs";
-import { makeInt } from "../generic-makers.mjs";
+import { E20 } from "../../../helpers/config.mjs";
+import { makeInt } from "../../generic-makers.mjs";
 
 const fields = foundry.data.fields;
